Refactor main script into smaller modules
